pyproject.toml
[tool.poetry]
name = "drf-starter"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "drf_starter"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
djangorestframework = "^3.14.0"
django-filter = "^23.2"
django-environ = "^0.10.0"
djangorestframework-simplejwt = "^5.2.2"
psycopg = {extras = ["binary", "pool"], version = "^3.1.9"}


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pytest-django = "^4.5.2"
factory-boy = "^3.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
